chore: move SendFormProps to place where its used
